requirements.txt
django
gunicorn
django-heroku
numpy
torch
djangorestframework